New ver 0.4.2
Config.content_provider now expects an instance of
LiquidAssets::Template to be returned.  The Template class's initializer
takes the template source and the modified time (defaults to Time.now).
This is to take advantage of the new cache in Rails 4.  The new cache
will check the mtime if present to avoid recompiling templates
unnecessarily.